Move IndexReader type beside IndexEntryProcessor
More consistent that way
