+++ pkg/chunk/storage_client.go
@@ -88,14 +88,3 @@ type StorageObject struct {
 // StorageCommonPrefix represents a common prefix aka a synthetic directory in Object Store.
 // It is guaranteed to always end with delimiter passed to List method.
 type StorageCommonPrefix string
-
-// IndexEntryProcessor receives index entries from the table.
-type IndexEntryProcessor interface {
-	ProcessIndexEntry(indexEntry IndexEntry) error
-
-	// Will this user be accepted by the processor?
-	AcceptUser(user string) bool
-
-	// Called at the end of reading of index entries.
-	Flush() error
-}
